# Conflicts:
#	Bundles/CompanyUnitAddressGui/src/Spryker/Zed/CompanyUnitAddressGui/Communication/Table/CompanyUnitAddressTable.php

+++ Bundles/ProductReviewSearch/src/Spryker/Zed/ProductReviewSearch/Communication/Plugin/PageMapExpander/ProductReviewMapExpanderPlugin.php
@@ -15,10 +15,13 @@ use Spryker\Zed\Search\Business\Model\Elasticsearch\DataMapper\PageMapBuilderInt
 
 /**
  * @method \Spryker\Zed\ProductReviewSearch\Communication\ProductReviewSearchCommunicationFactory getFactory()
+ * @method \Spryker\Zed\ProductReviewSearch\Business\ProductReviewSearchFacadeInterface getFacade()
  */
 class ProductReviewMapExpanderPlugin extends AbstractPlugin implements ProductPageMapExpanderInterface
 {
     /**
+     * @api
+     *
      * @param \Generated\Shared\Transfer\PageMapTransfer $pageMapTransfer
      * @param \Spryker\Zed\Search\Business\Model\Elasticsearch\DataMapper\PageMapBuilderInterface $pageMapBuilder
      * @param array $productData
@@ -41,8 +44,8 @@ class ProductReviewMapExpanderPlugin extends AbstractPlugin implements ProductPa
         $pageMapBuilder
             ->addSearchResultData($pageMapTransfer, 'rating', $rating)
             ->addSearchResultData($pageMapTransfer, 'review_count', $productData['review_count'])
-            ->addIntegerFacet($pageMapTransfer, 'rating', $rating * 100)
-            ->addIntegerSort($pageMapTransfer, 'rating', $rating * 100);
+            ->addIntegerFacet($pageMapTransfer, 'rating', (int)($rating * 100))
+            ->addIntegerSort($pageMapTransfer, 'rating', (int)($rating * 100));
 
         return $pageMapTransfer;
     }
